Renamed folders, installed clsx package
