+++ src/components/button/button.scss
@@ -1,5 +1,5 @@
-@import '../../stylesUniversal/mixins/mixins.scss';
-@import '../../stylesUniversal/variables/variables.scss';
+@import '../../styles/mixins/mixins.scss';
+@import '../../styles/variables/variables.scss';
 
 .btn {
     @include BTN ($white, 600, $main-color, none, 10px, flex, row, center, center);
