+++ src/components/footer/Footer.jsx
@@ -1,6 +1,7 @@
 import React from "react";
 import { NavLink } from "react-router-dom";
-import "./Footer.scss";
+
+import "./footer.scss";
 
 const Footer = () => {
   return (
