+++ src/components/footer/footer.scss
@@ -1,4 +1,4 @@
-@import '../../stylesUniversal/variables/variables.scss';
+@import '../../styles/variables/variables.scss';
 
 .footer {
   font-family: "Poppins-Light", "sans-serif";
