+++ src/components/header/Header.jsx
@@ -1,6 +1,7 @@
 import React from "react";
 import { NavLink } from "react-router-dom";
-import "./Header.scss";
+
+import "./header.scss";
 import logo from "../../assets/images/logo.png";
 import Button from "../button/Button";
 
@@ -37,7 +38,7 @@ const Header = () => {
                 </NavLink>
               </li>
               <li class="menu__item">
-                <Button className="header__button" text="Apply" />
+                <Button className="btn header__button" text="Apply" />
               </li>
             </ul>
           </nav>
