+++ src/components/header/header.scss
@@ -1,4 +1,4 @@
-@import '../../stylesUniversal/mixins/mixins.scss';
+@import '../../styles/mixins/mixins.scss';
 
 .header {
   font-family: 'Baloo2-Regular', sans-serif;
