--- src/componentsUniversal/button/button.js
@@ -1,9 +1,0 @@
-import './Button.scss'
-
-const Button = ({text, className='', onClick}) => {
-    return (
-        <button  className={`btn ${className}`} onClick={onClick}>{text}</button>
-    )
-}
-
-export default Button;
