+++ src/pages/home/homeSections/homeInfo/homeInfo.scss
