+++ src/pages/home/homeSections/homeSubscribe/homeSubscribe.scss
@@ -1,4 +1,4 @@
-@import '../../../../stylesUniversal/variables/variables.scss';
+@import '../../../../styles/variables/variables.scss';
 
 .subscribe {
   font-family: "Poppins-Regular", "sans-serif";
