+++ src/pages/modal/Modal.scss
@@ -1,4 +1,4 @@
-@import '../../stylesUniversal/variables/variables.scss';
+@import '../../styles/variables/variables.scss';
 
 .modal {
   position: fixed;
